ADDED SYSTEM DIAGRAMS
I added a few diagrams to the repo. This include the file for the top
level diagram, component diagrams (almost) for the Tutorial & Profile
sections, and some use case diagrams for those sections as well.

The .uxf file format is for UMLet which is availible in Windows. I'm
just using it as a matter of convience but I have found that it is
actually really good for making professional, quick diagrams.