Quick Design Mockup
Added a quick design mock up for the site. It's in no way functional nor
descriptive of the final design. I just wanted to add this so that I can
show at least some of the things I was think for the design. This is
kind of the front page. I used Adobe Muse to make this mock up (which is
pretty inflexible) but it gave me some ideas for how we can make the
site look at least.