Updates
I'm pretty sure that the rating stuff is really screwed up, so don't
look at that to much. I can't upload them to the main page to test how
they are working with the database so I don't know how any of the pages
will work. But I know for a fact that the rating won't work I'm pretty
sure I'll fix it tomorrow and Editing a tutorial is coming.